Hacking in a few IP Options.
Changed congestionLevel from int to double.
Added tags for IP Option RouteRecord and Strict source routing.
Trying to hack-in RouteRecord on all TCP-Data and Strict Source routing
on Acks. (Need to propagate the RouteRecord to TCP and copy it to the
Ack as Strict source).

+++ src/inet/routing/ospfv2/router/OspfRoutingTableEntry.cc
@@ -211,7 +211,7 @@ InterfaceEntry *RoutingTableEntry::getInterface() const
   if(!ift){
     return Ipv4Route::getInterface();
   }
-  int congestLevel = INT_MAX;
+  double congestLevel = INT_MAX;
   Ipv4Address nextHopAddr = Ipv4Address::UNSPECIFIED_ADDRESS;
   NextHop nextHop, tmpNextHop;
   nextHop.hopAddress = Ipv4Address::UNSPECIFIED_ADDRESS;
